S3 migration development started

+++ frmMain.cs
@@ -28,7 +28,7 @@ namespace CloudinaryMigration
         //    "336625465193458",
         //    "WuvXWKKtWJrcfNulItmGqvViSGU");
 
-        Account account = new Account("","","");
+        Account account = new Account("", "", "");
 
         Cloudinary cloudinary;
         SQLLiteHelper sQLLiteHelper;
@@ -36,23 +36,33 @@ namespace CloudinaryMigration
         static String tempPathForLongFilepaths = "";
         public frmMain()
         {
-            InitializeComponent();
-            sqlitedbpath = ConfigurationManager.AppSettings["sqllitedbpath"];
+            try
+            {
+                InitializeComponent();
+                sqlitedbpath = ConfigurationManager.AppSettings["sqllitedbpath"];
 
-            string cldAccountname = ConfigurationManager.AppSettings["cldAccountname"];
-            string apiKey = ConfigurationManager.AppSettings["apiKey"];
-            string secretKey = ConfigurationManager.AppSettings["secretKey"];
-            threadCount =Int16.Parse( ConfigurationManager.AppSettings["threadcount"].ToString());
-            tempPathForLongFilepaths = ConfigurationManager.AppSettings["tempPathForLongFilepaths"];
-            account = new Account(cldAccountname,apiKey,secretKey);
-            this.Text = "Account Name:" + cldAccountname + "; apiKey" + apiKey + "; secretKey" + secretKey;
+                string cldAccountname = ConfigurationManager.AppSettings["cldAccountname"];
+                string apiKey = ConfigurationManager.AppSettings["apiKey"];
+                string secretKey = ConfigurationManager.AppSettings["secretKey"];
+                threadCount = Int16.Parse(ConfigurationManager.AppSettings["threadcount"].ToString());
+                tempPathForLongFilepaths = ConfigurationManager.AppSettings["tempPathForLongFilepaths"];
+                account = new Account(cldAccountname, apiKey, secretKey);
+                this.Text = Application.ProductVersion + "; Account Name:" + cldAccountname + "; apiKey" + apiKey + "; secretKey" + secretKey;
 
-            sQLLiteHelper = new SQLLiteHelper(sqlitedbpath);
+                sQLLiteHelper = new SQLLiteHelper(sqlitedbpath);
 
-            Directory.Delete(tempPathForLongFilepaths, true);
-            Directory.CreateDirectory(tempPathForLongFilepaths);
+                if (Directory.Exists(tempPathForLongFilepaths))
+                    Directory.Delete(tempPathForLongFilepaths, true);
 
-            rchTextboxUnprocessedDtls.Text = "";
+                Directory.CreateDirectory(tempPathForLongFilepaths);
+
+                rchTextboxUnprocessedDtls.Text = "";
+
+            }
+            catch (Exception ex)
+            {
+                throw ex;
+            }
             //try
             //{
             //    string Path = "E:\\New folder\\testcld\\OneDrive - Global Eagle\\MARCH CYCLE (AWS WORKAROUND)\\AIRLINES\\AIR NEW ZEALAND\\0222\\Featured Images\\Crew Space Highlight Titles\\Series\\One Lane Bridge_69443\\One Lane Bridge_Season_02_93575\\Key Art\\OLB Digital Billboards Folder\\maosx\\test123456\\test.jpg";
@@ -152,9 +162,9 @@ namespace CloudinaryMigration
                     return;
                 }
                 Int64 rowIDfrom = sQLLiteHelper.getRowID(txtFolderSelected.Text.Trim(), txtDestinationpath.Text.Trim(), "min"); ;
-                Int64 rowIDto = sQLLiteHelper.getRowID(txtFolderSelected.Text.Trim(),txtDestinationpath.Text.Trim(),"max");
+                Int64 rowIDto = sQLLiteHelper.getRowID(txtFolderSelected.Text.Trim(), txtDestinationpath.Text.Trim(), "max");
 
-                if( (txtrowidfrom.Text.Trim().Length > 0 && Int64.Parse(txtrowidfrom.Text.Trim()) < rowIDfrom) ||
+                if ((txtrowidfrom.Text.Trim().Length > 0 && Int64.Parse(txtrowidfrom.Text.Trim()) < rowIDfrom) ||
                     (txtrowidto.Text.Trim().Length > 0 && Int64.Parse(txtrowidto.Text.Trim()) > rowIDto))
                 {
                     MessageBox.Show("Invalid Rowid From and To value, it should be between " + rowIDfrom + " and " + rowIDto);
@@ -162,18 +172,18 @@ namespace CloudinaryMigration
                 }
 
 
-                if (txtrowidfrom.Text.Trim().Length>0)
-                rowIDfrom = Int64.Parse(txtrowidfrom.Text.Trim());
+                if (txtrowidfrom.Text.Trim().Length > 0)
+                    rowIDfrom = Int64.Parse(txtrowidfrom.Text.Trim());
 
                 if (txtrowidto.Text.Trim().Length > 0)
-                rowIDto = Int64.Parse(txtrowidto.Text.Trim());
+                    rowIDto = Int64.Parse(txtrowidto.Text.Trim());
 
                 Int64 counter = 0;
                 String rowids = "";
-                while (rowIDfrom <= rowIDto )
+                while (rowIDfrom <= rowIDto)
                 {
                     String publicIDPath = "";
-                    App_log_dtls app_Log_Dtls= sQLLiteHelper.getApp_log_dtls(rowIDfrom);
+                    App_log_dtls app_Log_Dtls = sQLLiteHelper.getApp_log_dtls(rowIDfrom);
 
                     try
                     {
@@ -196,6 +206,7 @@ namespace CloudinaryMigration
                         }
 
                         String filepath = app_Log_Dtls.filename;
+
                         if (chkUploadPath.Checked)
                         {
                             publicIDPath = (txtDestinationpath.Text.Replace("\\", "/") + "/" + txtFolderSelected.Text + "/" + filepath.Replace(textBox1.Text + "\\", "")).Replace("\\", "/");
@@ -221,29 +232,30 @@ namespace CloudinaryMigration
                             Task.WaitAll(tasks.ToArray());
                             uploadDetails.Clear();
                         }
-                        
+
 
                     }
-                    catch(Exception ex)
+                    catch (Exception ex)
                     {
                         sQLLiteHelper.InsertData("Error", "Error while processing of rowid " + app_Log_Dtls.rowid, jobID, "");
                     }
-                    finally {
+                    finally
+                    {
                         rowIDfrom = rowIDfrom + 1;
                     }
 
                 }
-                
+
                 if (uploadDetails.Count > 0)
                 {
                     List<Task> tasks = new List<Task>();
 
                     foreach (var uploadfiledtl in uploadDetails)
                     {
-                        Task task = Task.Factory.StartNew(() => createUploadThread(uploadfiledtl, sQLLiteHelper, cloudinary, jobID,uploadfiledtl.rowid));
+                        Task task = Task.Factory.StartNew(() => createUploadThread(uploadfiledtl, sQLLiteHelper, cloudinary, jobID, uploadfiledtl.rowid));
                         tasks.Add(task);
                     }
-                    Task.WaitAll(tasks.ToArray());
+                    Task.WaitAll(tasks.ToArray(), 60000);
                     uploadDetails.Clear();
 
                 }
@@ -264,40 +276,72 @@ namespace CloudinaryMigration
 
         }
 
-        static void createUploadThread(UploadModel uploadfiledtl, SQLLiteHelper sQLLiteHelper, Cloudinary cloudinary, String jobID,Int64 rowid)
+        static void createUploadThread(UploadModel uploadfiledtl, SQLLiteHelper sQLLiteHelper, Cloudinary cloudinary, String jobID, Int64 rowid)
         {
 
             try
             {
                 String filePath = uploadfiledtl.FilePath;
 
-                if (uploadfiledtl.FilePath.Length>250) {
-                    //tempPathForLongFilepaths
-                    if (Alphaleonis.Win32.Filesystem.File.Exists(uploadfiledtl.FilePath)) { 
-                    }
+                if (uploadfiledtl.FilePath.Length > 250)
+                {
 
-                    
                     Byte[] filebytes = Alphaleonis.Win32.Filesystem.File.ReadAllBytes(filePath);
 
-                    
                     String newFileNamewithPath = tempPathForLongFilepaths + "\\" + uploadfiledtl.rowid + "_" + Alphaleonis.Win32.Filesystem.Path.GetFileName(filePath);
 
-                    File.WriteAllBytes(newFileNamewithPath , filebytes);
+                    File.WriteAllBytes(newFileNamewithPath, filebytes);
                     filePath = newFileNamewithPath;
-                    //System.IO.File.Copy(Path, "D:\\abc.jpg");
 
                 }
 
-                var uploadParams = new ImageUploadParams()
+                Int64 size = new FileInfo(filePath).Length / 1024;
+                if ((size / 1024) > 100)
                 {
-                    File = new FileDescription(filePath),
-                    PublicId = uploadfiledtl.publicID,
+                    var uploadParams = new RawUploadParams()
+                    {
+                        File = new FileDescription(filePath),
+                        PublicId = uploadfiledtl.publicID.Substring(0, uploadfiledtl.publicID.LastIndexOf(".")),
 
-                };
+                    };
 
-                var uploadResult = cloudinary.Upload(uploadParams, "auto");
-                sQLLiteHelper.InsertData("Info", rowid.ToString() + " | Upload completed", jobID, uploadfiledtl.publicID);
-                sQLLiteHelper.updateRowIDStatus(rowid, 1, 0, "");
+                    var uploadResult = cloudinary.UploadLarge(uploadParams);
+                    if (uploadResult.Error == null)
+                    {
+                        sQLLiteHelper.InsertData("Info", rowid.ToString() + " | Upload completed", jobID, uploadfiledtl.publicID);
+                        sQLLiteHelper.updateRowIDStatus(rowid, 1, 0, "");
+
+                    }
+                    else
+                    {
+                        sQLLiteHelper.InsertData("Info", rowid.ToString() + " | Upload Error" + uploadResult.Error.Message, jobID, uploadfiledtl.publicID);
+                        sQLLiteHelper.updateRowIDStatus(rowid, 0, 0, uploadResult.Error.Message);
+                    }
+
+                }
+                else
+                {
+                    var uploadParams = new ImageUploadParams()
+                    {
+                        File = new FileDescription(filePath),
+                        PublicId = uploadfiledtl.publicID.Substring(0, uploadfiledtl.publicID.LastIndexOf(".")) + "a001",
+
+                    };
+
+                    var uploadResult = cloudinary.Upload(uploadParams, "auto");
+                    if (uploadResult.Error == null)
+                    {
+                        sQLLiteHelper.InsertData("Info", rowid.ToString() + " | Upload completed", jobID, uploadfiledtl.publicID);
+                        sQLLiteHelper.updateRowIDStatus(rowid, 1, 0, "");
+
+                    }
+                    else
+                    {
+                        sQLLiteHelper.InsertData("Info", rowid.ToString() + " | Upload Error" + uploadResult.Error.Message, jobID, uploadfiledtl.publicID);
+                        sQLLiteHelper.updateRowIDStatus(rowid, 0, 0, uploadResult.Error.Message);
+                    }
+
+                }
 
             }
             catch (Exception ex)
@@ -310,26 +354,51 @@ namespace CloudinaryMigration
 
         private void button1_Click(object sender, EventArgs e)
         {
-
-            if (sQLLiteHelper.getRowID(txtFolderSelected.Text.Trim(), txtDestinationpath.Text.Trim(),"max")>0) {
-                MessageBox.Show("folder and DB is already synced");
-                return;
-            }
-
-            if (txtDestinationpath.Text.Contains("\\"))
+            Boolean checkfilenameindb = false;
+            if (sQLLiteHelper.getRowID(txtFolderSelected.Text.Trim(), txtDestinationpath.Text.Trim(), "max") > 0)
             {
-                MessageBox.Show("Invalid character in Cloudinary path", "", MessageBoxButtons.OK);
-                return;
+                checkfilenameindb = true;
+                DialogResult dgResult = MessageBox.Show("folder and DB is already synced", "migration", MessageBoxButtons.YesNo,MessageBoxIcon.Information);
+                if (dgResult == DialogResult.No)
+                {
+                    return;
+                }
             }
-
-
-            foreach (var filepath in Directory.EnumerateFiles(textBox1.Text, "*", SearchOption.AllDirectories))
+            try
             {
+                if (txtDestinationpath.Text.Contains("\\"))
+                {
+                    MessageBox.Show("Invalid character in Cloudinary path", "", MessageBoxButtons.OK);
+                    return;
+                }
 
-                sQLLiteHelper.InsertDataFordbSync(txtFolderSelected.Text.Trim(), filepath, txtDestinationpath.Text.Trim(), 0, 0, "");
+
+                foreach (var filepath in Directory.EnumerateFiles(textBox1.Text, "*", SearchOption.AllDirectories))
+                {
+                    if (checkfilenameindb)
+                    {
+                        if (sQLLiteHelper.getRowID(txtFolderSelected.Text.Trim(), txtDestinationpath.Text.Trim(), "count", filepath) == 0)
+                        {
+                            sQLLiteHelper.InsertDataFordbSync(txtFolderSelected.Text.Trim(), filepath, txtDestinationpath.Text.Trim(), 0, 0, "");
+                        }
+
+                    }
+                    else
+                    {
+                        sQLLiteHelper.InsertDataFordbSync(txtFolderSelected.Text.Trim(), filepath, txtDestinationpath.Text.Trim(), 0, 0, "");
+                    }
+
+
+                }
+
+                MessageBox.Show("synched to DB!");
 
             }
-            MessageBox.Show("synched to DB!");
+            catch (Exception ex)
+            {
+                MessageBox.Show("Error while sync " + ex.Message, "Migration");
+            }
+
         }
 
         private void textBox2_TextChanged(object sender, EventArgs e)
